refactoring review.js
