+++ review.html
@@ -21,7 +21,7 @@
   </div>
 
   <!-- initialize your application -->
-  <script type="module" src="./src/init/review.js"></script>
+  <script type="module" src="./src/review.js"></script>
 
   <!-- load the review buttons for your live deployment  -->
   <script type="module" src="./lib/review-buttons.js"></script>
