+++ src/handlers/goToindexHandler.js
@@ -1,0 +1,5 @@
+'use strict';
+
+export const goToIndex = () => {
+    window.location.replace("./index.html");
+}
