+++ src/listeners/goToIndex.js
@@ -1,0 +1,7 @@
+'use strict';
+
+import { goToIndex } from "../handlers/goToindexHandler.js";
+
+const btnIndex = document.getElementById('retakeQuiz');
+
+btnIndex.addEventListener('click', goToIndex);
